@@ -7,6 +7,7 @@ import (
 
 	"github.com/apex/log"
 	"github.com/docker/docker/api/types/container"
+
 	"github.com/pterodactyl/wings/config"
 )
 
@@ -25,7 +26,7 @@ type Mount struct {
 	// that we're mounting into the container at the Target location.
 	Source string `json:"source"`
 
-	// Whether or not the directory is being mounted as read-only. It is up to the environment to
+	// Whether the directory is being mounted as read-only. It is up to the environment to
 	// handle this value correctly and ensure security expectations are met with its usage.
 	ReadOnly bool `json:"read_only"`
 }
