@@ -6,12 +6,10 @@ import (
 	"os"
 
 	"emperror.dev/errors"
-
-	"github.com/pterodactyl/wings/server/filesystem"
-
 	"github.com/mholt/archiver/v3"
 
 	"github.com/pterodactyl/wings/remote"
+	"github.com/pterodactyl/wings/server/filesystem"
 )
 
 type LocalBackup struct {
@@ -90,7 +88,7 @@ func (b *LocalBackup) Restore(ctx context.Context, _ io.Reader, callback Restore
 			if f.IsDir() {
 				return nil
 			}
-			return callback(filesystem.ExtractNameFromArchive(f), f, f.Mode())
+			return callback(filesystem.ExtractNameFromArchive(f), f, f.Mode(), f.ModTime(), f.ModTime())
 		}
 	})
 }
