@@ -1,6 +1,3 @@
 package system
 
-var (
-	// The current version of this software.
-	Version = "0.0.1"
-)
+var Version = "develop"
